# LibGfx/WebPWriterLossless: Use predictor transform