@nico 2024-07-27
I should mention that the QOI writer is 225 lines, while the WebP writer is 771 lines (…and the WebP writer uses LibCompress to do some of the huffman coding). QOI has a very good simplicity / compression tradeoff.