@nico 2024-08-21
Similar to https://github.com/SerenityOS/serenity/pull/24819#issuecomment-2301975550 , the numbers were wrong due to me not passing -A (show apparent size) to du. With that:

```
% du -hAs webp-out-subtract-green-only 
 71M	webp-out-subtract-green-only
% du -hAs webp-out-afterer
 53M	webp-out-afterer
```

So instead of "from 88M to 69M (21.6%)", it's actually from 71M to 53M (25.3%).